Add files via upload
In this project, we'll use OPENGL to depict the scenario of a wet day on a city road. At first, traffic goes day and night on the road. We'll try to depict the rain falling in the setting of a rainy day. Additionally, we'll demonstrate an animation of a daytime sunset, moving traffic, and train motion. An example of a whirling windmill